// src/modules/textile/reports/knitting/grey-fabric-issue-challan-report/grey-fabric-issue-challan-report-type.ts
export type IGreyFabricIssueChallanReport = {
  DTLSID: number;
  INTER_COMPANY_ID: number;
  COMPANY_NAME: string;
  COMPANY_PREFIX: string;
  COMPANY_ADDRESS: string;
  COMPANY_CONTACT: string;
  COMPANY_EMAIL: string;
  COMPANY_WEB: string;
  COMPANY_REMARKS: string;
  PARTY_ID: number;
  PARTY: string;
  PARTY_MOBILE: string;
  PARTYPHONE: string;
  PARTY_ADDRESS: string;
  PARTY_FAX: string;
  PARTY_EMAIL: string;
  PARTY_WEB: string;
  RCV_CHALLAN_NO: string;
  YARN_CHALLAN_ID: number;
  CHALLAN_NO: string;
  CHALLAN_DATE: Date;
  BUYER_NAME: string;
  BUYER_ID: number;
  PO_NO: string;
  PO_ID: number;
  STYLE: string;
  CREATED_BY: string;
  CREATED_BY_NAME: string;
  CREATED_DATE: Date;
  SORTINGNO: number;
  GMT_GSM: string;
  FABRIC_ID: number;
  FABRIC: string;
  YARN_BRAND: string;
  YARN_LOT_NUMBER: string;
  GMT_COLOR: string;
  GMT_SIZE: string;
  YARN_CHALLAN_REQ_DIA: string;
  ROLL_QTY: number;
  QTY_KG: number;
  QTY_PCS: number;
  GREY_SHAPE: string;
  UOM: string;
  DTLS_REMARKS: string;
  MST_REMARKS: string;
  TOTAL_LOSE_YARN: number;
  GAUGE: string;
  STITCH_LENGTH: string;
  ISSUE_NAME: string;
  YARN_COUNT: string;
  BUYER: string;
  PO: string;
};